requirements.txt
openai
httpx 
tqdm
google-generativeai
# dotenv
# pip install -q -U google-generativeai